[Fix] 지우개 감도 조정, 하이라이트 드로잉에만 지우게 수정

+++ Project/Reazy/Views/PDF/Drawing/PDFDrawer.swift
@@ -70,7 +70,7 @@ class PDFDrawer {
 }
 
 extension PDFDrawer: DrawingGestureRecognizerDelegate {
-    // 모든 드로잉 꺼내서 pdfview에 페이지 별로 추가하는 함수
+    // MARK: -모든 드로잉 꺼내서 pdfview에 페이지 별로 추가하는 함수
     func loadDrawings() {
         pdfView.goToFirstPage(nil)
         
@@ -86,7 +86,7 @@ extension PDFDrawer: DrawingGestureRecognizerDelegate {
         pdfView.setNeedsDisplay()
     }
     
-    // 제스처 최초 시작 시 한 번 실행되는 함수
+    // MARK: -제스처 최초 시작 시 한 번 실행되는 함수
     func gestureRecognizerBegan(_ location: CGPoint) {
         guard let page = pdfView.page(for: location, nearest: true) else { return }
         currentPage = page
@@ -100,7 +100,7 @@ extension PDFDrawer: DrawingGestureRecognizerDelegate {
         }
     }
     
-    // 제스처 움직이는 동안 실행되는 함수
+    // MARK: -제스처 움직이는 동안 실행되는 함수
     func gestureRecognizerMoved(_ location: CGPoint) {
         guard let page = currentPage else { return }
         let convertedPoint = pdfView.convert(location, to: page)
@@ -135,6 +135,7 @@ extension PDFDrawer: DrawingGestureRecognizerDelegate {
     }
     
     private func updateEraserLayer(at location: CGPoint) {
+        // 지우개 레이어가 없다면 새로 생성
         if eraserLayer == nil {
             eraserLayer = CAShapeLayer()
             eraserLayer?.fillColor = UIColor(hex: "#EFEFF8").cgColor
@@ -143,11 +144,14 @@ extension PDFDrawer: DrawingGestureRecognizerDelegate {
             pdfView.layer.addSublayer(eraserLayer!)
         }
         
-        let eraserCirclePath = UIBezierPath(arcCenter: location, radius: 14, startAngle: 0, endAngle: .pi * 2, clockwise: true)
+        // 확대 축소 비율 고려해서 지우개 모양 만들기
+        let scaleFactor = pdfView.scaleFactor
+        let eraserRadius = 5 * scaleFactor
+        let eraserCirclePath = UIBezierPath(arcCenter: location, radius: eraserRadius, startAngle: 0, endAngle: .pi * 2, clockwise: true)
         eraserLayer?.path = eraserCirclePath.cgPath
     }
     
-    // 패드에서 제스처 뗄 때 실행되는 함수
+    // MARK: - 패드에서 제스처 뗄 때 실행되는 함수
     func gestureRecognizerEnded(_ location: CGPoint) {
         guard let page = currentPage else { return }
         let convertedPoint = pdfView.convert(location, to: page)
@@ -192,15 +196,15 @@ extension PDFDrawer: DrawingGestureRecognizerDelegate {
         forceRedraw(annotation: currentAnnotation!, onPage: onPage)
     }
     
-    // 획을 그리고 배열에 저장하는 함수
+    // MARK: -획을 그리고 배열에 저장하는 함수
     private func createFinalAnnotation(path: UIBezierPath, page: PDFPage) -> PDFAnnotation {
         let border = PDFBorder()
         border.lineWidth = drawingTool.width
         
         let bounds = CGRect(x: path.bounds.origin.x - 5,
-                            y: path.bounds.origin.y - 5,
-                            width: path.bounds.size.width + 10,
-                            height: path.bounds.size.height + 10)
+                              y: path.bounds.origin.y - 5,
+                              width: path.bounds.size.width + 10,
+                              height: path.bounds.size.height + 10)
         let signingPathCentered = UIBezierPath()
         signingPathCentered.cgPath = path.cgPath
         let _ = signingPathCentered.moveCenter(to: bounds.center)
@@ -222,10 +226,12 @@ extension PDFDrawer: DrawingGestureRecognizerDelegate {
         return annotation
     }
     
-    // 지우개로 주석 지울 때 실행되는 함수
+    // MARK: - 지우개로 주석 지울 때 실행되는 함수
     private func removeAnnotationAtPoint(point: CGPoint, page: PDFPage) {
         let convertedPoint = pdfView.convert(point, to: page)
-        let hitTestRect = CGRect(x: convertedPoint.x - 3, y: convertedPoint.y - 3, width: 6, height: 6)
+        let scaleFactor = pdfView.scaleFactor
+        let scaledRadius = 5 / scaleFactor
+        let hitTestRect = CGRect(x: convertedPoint.x - scaledRadius, y: convertedPoint.y - scaledRadius, width: scaledRadius * 2, height: scaledRadius * 2)
         
         let annotations = page.annotations.filter { annotation in
             annotation.bounds.intersects(hitTestRect)
@@ -233,23 +239,25 @@ extension PDFDrawer: DrawingGestureRecognizerDelegate {
         
         if let pageIndex = pdfView.document?.index(for: page) {
             for annotation in annotations {
-                let annotationBounds = annotation.bounds
-                
-                let indicesToRemove = drawingDataArray.indices.filter { index in
-                    let drawing = drawingDataArray[index]
-                    return drawing.pageIndex == pageIndex && drawing.path.bounds.intersects(annotationBounds)
-                }
-                
-                // 뒤쪽 인덱스 주석부터 제거
-                for index in indicesToRemove.reversed() {
-                    let drawing = drawingDataArray[index]
-                    let id = drawing.id
+                // 하이라이트랑 드로잉만 지우기
+                if annotation.type == "Ink" || (annotation.type == "Highlight" && annotation.value(forAnnotationKey: .contents) == nil) {
+                    let annotationBounds = annotation.bounds
                     
-                    _ = drawingService.deleteDrawingData(for: pdfID, id: id)
-                    drawingDataArray.remove(at: index)
+                    let indicesToRemove = drawingDataArray.indices.filter { index in
+                        let drawing = drawingDataArray[index]
+                        return drawing.pageIndex == pageIndex && drawing.path.bounds.intersects(annotationBounds)
+                    }
+                    
+                    // 겹쳐진 부분 위층부터 지우기 위해 뒤쪽 인덱스 주석부터 제거
+                    for index in indicesToRemove.reversed() {
+                        let drawing = drawingDataArray[index]
+                        let id = drawing.id
+                        
+                        _ = drawingService.deleteDrawingData(for: pdfID, id: id)
+                        drawingDataArray.remove(at: index)
+                    }
+                    page.removeAnnotation(annotation)
                 }
-                
-                page.removeAnnotation(annotation)
             }
         }
     }
